requirements.txt
requests
rich
python-dotenv
tzdata
motor
pymongo
passlib[bcrypt]
PyJWT
python-multipart
